+++ packages/health/lib/src/data_types.dart
@@ -48,6 +48,7 @@ enum HealthDataAccess {
   WRITE,
   READ_WRITE,
 }
+
 /// List of data types available on iOS
 const List<HealthDataType> _dataTypeKeysIOS = [
   HealthDataType.ACTIVE_ENERGY_BURNED,